requirements.txt
oci
oci[adk]
requests
fdk
oracledb
pandas
python-dotenv
